chore: Reorganize provider configurations
