+++ src/app/src/pages/redteam/setup/components/Targets/index.tsx
@@ -22,6 +22,7 @@ import type { ProviderOptions } from '../../types';
 import Prompts from '../Prompts';
 import { predefinedTargets, customTargetOption } from '../constants';
 import BrowserAutomationConfiguration from './BrowserAutomationConfiguration';
+import CommonConfigurationOptions from './CommonConfigurationOptions';
 import CustomTargetConfiguration from './CustomTargetConfiguration';
 import HttpEndpointConfiguration from './HttpEndpointConfiguration';
 import TestTargetConfiguration from './TestTargetConfiguration';
@@ -53,10 +54,6 @@ const validateUrl = (url: string, type: 'http' | 'websocket' = 'http'): boolean 
   }
 };
 
-const requiresTransformResponse = (target: ProviderOptions) => {
-  return target.id === 'http' || target.id === 'websocket';
-};
-
 const requiresPrompt = (target: ProviderOptions) => {
   return target.id !== 'http' && target.id !== 'websocket' && target.id !== 'browser';
 };
@@ -77,7 +74,7 @@ export default function Targets({ onNext, onBack, setupModalOpen }: TargetsProps
     suggestions?: string[];
     providerResponse?: ProviderResponse;
   } | null>(null);
-  const [hasTestedTarget, setHasTestedTarget] = useState(false);
+
   const [bodyError, setBodyError] = useState<string | null>(null);
   const [urlError, setUrlError] = useState<string | null>(null);
   const [missingFields, setMissingFields] = useState<string[]>([]);
@@ -318,7 +315,6 @@ export default function Targets({ onNext, onBack, setupModalOpen }: TargetsProps
           message: 'Target configuration is valid!',
           providerResponse: data.providerResponse,
         });
-        setHasTestedTarget(true);
       }
     } catch (error) {
       console.error('Error testing target:', error);
@@ -476,15 +472,20 @@ export default function Targets({ onNext, onBack, setupModalOpen }: TargetsProps
         )}
       </Box>
 
+      <Typography variant="h6" gutterBottom>
+        Common Configuration
+      </Typography>
+      <CommonConfigurationOptions
+        selectedTarget={selectedTarget}
+        updateCustomTarget={updateCustomTarget}
+      />
+
       {testingEnabled && (
         <TestTargetConfiguration
           testingTarget={testingTarget}
           handleTestTarget={handleTestTarget}
           selectedTarget={selectedTarget}
           testResult={testResult}
-          requiresTransformResponse={requiresTransformResponse}
-          updateCustomTarget={updateCustomTarget}
-          hasTestedTarget={hasTestedTarget}
         />
       )}
